Logout component
Put logout component in sperate file to keep clean
